+++ cmd/mq_json/config.go
@@ -73,7 +73,7 @@ func initConfig() error {
 	flag.StringVar(&config.interval, "ibmmq.interval", "10", "How many seconds between each collection")
 
 	flag.BoolVar(&config.cc.ClientMode, "ibmmq.client", false, "Connect as MQ client")
-	flag.BoolVar(&config.qStatus, "ibmmq.qstatus", false, "Add QSTATUS data")
+	flag.BoolVar(&config.qStatus, "ibmmq.qStatus", false, "Add QSTATUS data")
 
 	flag.StringVar(&config.logLevel, "log.level", "error", "Log level - debug, info, error")
 	flag.StringVar(&config.pollInterval, "pollInterval", defaultPollInterval, "Frequency of checking channel status")